Conflicts:
	net.sf.redmine_mylyn.api/src/net/sf/redmine_mylyn/api/client/RedmineApiIssueProperty.java

+++ net.sf.redmine_mylyn.core/src/net/sf/redmine_mylyn/internal/core/client/ClientManager.java
@@ -19,6 +19,7 @@ import net.sf.redmine_mylyn.core.RedmineCorePlugin;
 import net.sf.redmine_mylyn.core.RedmineStatusException;
 import net.sf.redmine_mylyn.core.client.ClientFactory;
 import net.sf.redmine_mylyn.core.client.IClient;
+import net.sf.redmine_mylyn.internal.core.Messages;
 
 import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.Status;
@@ -28,7 +29,7 @@ import org.eclipse.mylyn.tasks.core.TaskRepositoryLocationFactory;
 
 public class ClientManager implements IRedmineClientManager {
 
-	private final static char[] ILLEGAL_ZIP_ENTRY_CHARS = "/*\\".toCharArray();
+	private final static char[] ILLEGAL_ZIP_ENTRY_CHARS = "/*\\".toCharArray(); //$NON-NLS-1$
 	
 	private Map<String, IClient> clientByUrl = new HashMap<String, IClient>();
 
@@ -119,7 +120,7 @@ public class ClientManager implements IRedmineClientManager {
 						String name = zipEntry.getName();
 						name = name.substring(0, name.length()-4); //.xml
 						for (char chr : ILLEGAL_ZIP_ENTRY_CHARS) {
-							name = name.replace("0x"+Integer.toHexString(chr), ""+chr);
+							name = name.replace("0x"+Integer.toHexString(chr), ""+chr); //$NON-NLS-1$ //$NON-NLS-2$
 						}
 						
 						confByUrl.put(name, Configuration.fromStream(zip));
@@ -131,7 +132,7 @@ public class ClientManager implements IRedmineClientManager {
 					}
 				}
 			} catch (Exception e) {
-				IStatus status = new Status(IStatus.ERROR, RedmineCorePlugin.PLUGIN_ID, "The Redmine respository data cache could not be read", e);
+				IStatus status = new Status(IStatus.ERROR, RedmineCorePlugin.PLUGIN_ID, Messages.ERRMSG_CANT_READ_CACHEDATA, e);
 				RedmineCorePlugin.getDefault().getLog().log(status);
 
 				confByUrl.clear();
@@ -149,7 +150,7 @@ public class ClientManager implements IRedmineClientManager {
 				}
 			} catch (Throwable e) {
 				StatusHandler.log(new Status(IStatus.WARNING, RedmineCorePlugin.PLUGIN_ID,
-						"The Redmine respository data cache could not be read", e));
+						Messages.ERRMSG_CANT_READ_CACHEDATA, e));
 			} finally {
 				if (in != null) {
 					try {
@@ -169,33 +170,37 @@ public class ClientManager implements IRedmineClientManager {
 				zipedCacheFile.delete();
 			}
 			
-			ZipOutputStream zip = null;
-			try {
-				try {
-					zip = new ZipOutputStream(new FileOutputStream(zipedCacheFile));
-					
-					for(Entry<String, Configuration>  entry : confByUrl.entrySet()) {
-						String name = entry.getKey();
-						for (char chr : ILLEGAL_ZIP_ENTRY_CHARS) {
-							name = name.replace(""+chr, "0x"+Integer.toHexString(chr));
-						}
-						zip.putNextEntry(new ZipEntry(name + ".xml"));
-						entry.getValue().write(zip);
-						zip.closeEntry();
-					}
-					
-				} finally {
-					if(zip!=null) {
-						zip.close();
-					}
-				}
-			} catch (Exception e) {
-				IStatus status = new Status(IStatus.ERROR, RedmineCorePlugin.PLUGIN_ID, "The Redmine respository data cache could not be written", e);
-				RedmineCorePlugin.getDefault().getLog().log(status);
+			if (confByUrl.size()>0) {
 				
-				if(zipedCacheFile.exists()) {
-					zipedCacheFile.delete();
+				ZipOutputStream zip = null;
+				try {
+					try {
+						zip = new ZipOutputStream(new FileOutputStream(zipedCacheFile));
+						
+						for(Entry<String, Configuration>  entry : confByUrl.entrySet()) {
+							String name = entry.getKey();
+							for (char chr : ILLEGAL_ZIP_ENTRY_CHARS) {
+								name = name.replace(""+chr, "0x"+Integer.toHexString(chr)); //$NON-NLS-1$ //$NON-NLS-2$
+							}
+							zip.putNextEntry(new ZipEntry(name + ".xml")); //$NON-NLS-1$
+							entry.getValue().write(zip);
+							zip.closeEntry();
+						}
+						
+					} finally {
+						if(zip!=null) {
+							zip.close();
+						}
+					}
+				} catch (Exception e) {
+					IStatus status = new Status(IStatus.ERROR, RedmineCorePlugin.PLUGIN_ID, Messages.ERRMSG_CANT_WRITE_CACHEDATA, e);
+					RedmineCorePlugin.getDefault().getLog().log(status);
+					
+					if(zipedCacheFile.exists()) {
+						zipedCacheFile.delete();
+					}
 				}
+				
 			}
 		}
 	}
